# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WatchFolderCreateRequest": {
        "properties": {
          "auto_run": {
            "default": true,
            "title": "Auto Run",
            "type": "boolean"
          },
          "cooldown_ms": {
            "default": 2000,
            "maximum": 60000.0,
            "minimum": 250.0,
            "title": "Cooldown Ms",
            "type": "integer"
          },
          "failed_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Folder"
          },
          "input_folder": {
            "title": "Input Folder",
            "type": "string"
          },
          "is_enabled": {
            "default": true,
            "title": "Is Enabled",
            "type": "boolean"
          },
          "move_failed_files": {
            "default": false,
            "title": "Move Failed Files",
            "type": "boolean"
          },
          "move_processed_files": {
            "default": false,
            "title": "Move Processed Files",
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "output_folder": {
            "title": "Output Folder",
            "type": "string"
          },
          "preset_id": {
            "title": "Preset Id",
            "type": "string"
          },
          "processed_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed Folder"
          },
          "skip_duplicates": {
            "default": true,
            "title": "Skip Duplicates",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "input_folder",
          "output_folder",
          "preset_id"
        ],
        "title": "WatchFolderCreateRequest",
        "type": "object"
      },
      "WatchFolderItem": {
        "properties": {
          "auto_run": {
            "default": true,
            "title": "Auto Run",
            "type": "boolean"
          },
          "cooldown_ms": {
            "default": 2000,
            "maximum": 60000.0,
            "minimum": 250.0,
            "title": "Cooldown Ms",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failed_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Folder"
          },
          "files_processed_count": {
            "default": 0,
            "title": "Files Processed Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_folder": {
            "title": "Input Folder",
            "type": "string"
          },
          "is_enabled": {
            "default": true,
            "title": "Is Enabled",
            "type": "boolean"
          },
          "last_activity_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_processed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Processed At"
          },
          "move_failed_files": {
            "default": false,
            "title": "Move Failed Files",
            "type": "boolean"
          },
          "move_processed_files": {
            "default": false,
            "title": "Move Processed Files",
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "output_folder": {
            "title": "Output Folder",
            "type": "string"
          },
          "preset_id": {
            "title": "Preset Id",
            "type": "string"
          },
          "preset_missing": {
            "default": false,
            "title": "Preset Missing",
            "type": "boolean"
          },
          "processed_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed Folder"
          },
          "skip_duplicates": {
            "default": true,
            "title": "Skip Duplicates",
            "type": "boolean"
          },
          "status": {
            "default": "idle",
            "enum": [
              "idle",
              "watching",
              "paused",
              "error",
              "unsupported"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "input_folder",
          "output_folder",
          "preset_id",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "WatchFolderItem",
        "type": "object"
      },
      "WatchFolderUpdateRequest": {
        "properties": {
          "auto_run": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Run"
          },
          "cooldown_ms": {
            "anyOf": [
              {
                "maximum": 60000.0,
                "minimum": 250.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldown Ms"
          },
          "failed_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Folder"
          },
          "input_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Folder"
          },
          "is_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Enabled"
          },
          "move_failed_files": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Move Failed Files"
          },
          "move_processed_files": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Move Processed Files"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "output_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Folder"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "processed_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed Folder"
          },
          "skip_duplicates": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Duplicates"
          }
        },
        "title": "WatchFolderUpdateRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ClearCut",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/watch-folders": {
      "get": {
        "operationId": "list_watch_folders_watch_folders_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchFolderItem"
                  },
                  "title": "Response List Watch Folders Watch Folders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Watch Folders",
        "tags": [
          "watch-folders"
        ]
      },
      "post": {
        "operationId": "create_watch_folder_watch_folders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchFolderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchFolderItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Watch Folder",
        "tags": [
          "watch-folders"
        ]
      }
    },
    "/watch-folders/{watch_folder_id}": {
      "delete": {
        "operationId": "delete_watch_folder_watch_folders__watch_folder_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "watch_folder_id",
            "required": true,
            "schema": {
              "title": "Watch Folder Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Watch Folder Watch Folders  Watch Folder Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watch Folder",
        "tags": [
          "watch-folders"
        ]
      },
      "patch": {
        "operationId": "update_watch_folder_watch_folders__watch_folder_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "watch_folder_id",
            "required": true,
            "schema": {
              "title": "Watch Folder Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchFolderUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchFolderItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Watch Folder",
        "tags": [
          "watch-folders"
        ]
      }
    },
    "/watch-folders/{watch_folder_id}/disable": {
      "post": {
        "operationId": "disable_watch_folder_watch_folders__watch_folder_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "watch_folder_id",
            "required": true,
            "schema": {
              "title": "Watch Folder Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchFolderItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Watch Folder",
        "tags": [
          "watch-folders"
        ]
      }
    },
    "/watch-folders/{watch_folder_id}/enable": {
      "post": {
        "operationId": "enable_watch_folder_watch_folders__watch_folder_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "watch_folder_id",
            "required": true,
            "schema": {
              "title": "Watch Folder Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchFolderItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Watch Folder",
        "tags": [
          "watch-folders"
        ]
      }
    },
    "/watch-folders/{watch_folder_id}/status": {
      "get": {
        "operationId": "watch_folder_status_watch_folders__watch_folder_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "watch_folder_id",
            "required": true,
            "schema": {
              "title": "Watch Folder Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchFolderItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Watch Folder Status",
        "tags": [
          "watch-folders"
        ]
      }
    }
  }
}